Creates tests for current navigation flow
