+++ app/build.gradle
@@ -42,11 +42,8 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 
-    //Mockito
-    testImplementation 'org.mockito:mockito-core:2.24.5'
-    androidTestImplementation 'org.mockito:mockito-android:2.24.5'
-
     // VERSION CATALOG
     implementation(libs.bundles.navigationx)
     debugImplementation(libs.bundles.fragmentx)
+    androidTestImplementation(libs.espresso.core)
 }
