+++ gradle/libs.versions.toml
@@ -3,6 +3,7 @@ nav_version = "2.5.3"
 fragment_test_version = "1.6.0-rc01"
 testx_version = "1.5.0"
 testx_runner_version = "1.5.2"
+espresso_version = "3.4.0"
 
 [libraries]
 nav-fragment = { group = "androidx.navigation", name = "navigation-fragment", version.ref = "nav_version" }
@@ -18,6 +19,7 @@ test-core = { group = "androidx.test", name = "core", version.ref = "testx_versi
 test-rules = { group = "androidx.test", name = "rules", version.ref = "testx_version" }
 test-runner = { group = "androidx.test", name = "runner", version.ref = "testx_runner_version" }
 
+espresso-core = { group = "androidx.test.espresso", name="espresso-core", version.ref="espresso_version"}
 [bundles]
 navigationx = ["nav-fragment", "nav-ui", "nav-fragment-kotlin", "nav-ui-kotlin", "nav-dynamic-features", "navigation-testing"]
 fragmentx = ["fragment-testing", "fragment-ktx", "test-core", "test-rules", "test-runner"]
